requirements.txt
streamlit==1.34.0
fpdf==1.7.2
Pillow==10.3.0
